+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Supported floating TOC feature for articles. (@MidnightCrowing)
 - Supported Miniflux format OPML import. (#768)
+- Supported displaying Youtube video in image gallery view.
 
 ### Changed